Azure Mobile Service Error
The app was breaking when an incorrect Mobile Service Address or
Application Key was given, even if enableAzureMobileService was set to
false. I just added an if-statement that checks to see if Azure is
enabled before trying to create the client. I also added in an async
message if Azure is enabled and the key/address is still incorrect.