Fixed RSS Service
If a bad RSS feed is added to the RssAddressCollection, any RSS feeds
following it would occasionally not be loaded. I moved the try-catch
statement in the RssService GetItems method to wrap just the part where
the feed is parsed. Before, if it broke on a bad feed, it would return
the results before parsing through the rest of the addresses.

I also fixed the Tumblr example, which is what caused me to find the bug
in the first place. (The blog I was pulling from decided to change their
address.)