Front End For XPCK
A front end that makes developing an XPCK app very simple. User enters
the various information and clicks the button, and then a new
application is generating with their information and saved in a new
folder named whatever they put in the "App Name" field. In addition to
the various settings, users can also select an image to use as their
logo, and it will be automatically resized & saved for both Windows
Phone and Windows 8.

This is still fairly new, so use at your own risk, and let me know if
you find any bugs!

I'll add ability to change theme colors in a future release.