Add Remote RSS fetch feature
From Roschler:
Note, I have left the AppSettings variables related to RSS and remote
RSS in the state necessary for easy testing. Reset them to what is best
for public consumption after you test this update and merge it into the
master.

This version adds remote RSS source list capability. There is a new
AppSettings variable called RemoteRssSourceUrl. If it is not empty and
EnableRssServce is TRUE, then the text file referenced by the URL will
be retrieved from the host server. The RSS feeds listed in that text
file will be added to the list in RssAddressCollection. This allows you
to add/remove RSS feeds dynamically to your app simply by altering the
contents of the publicly accessible web file. When run from the debugger
a date/time argument is added to the URL contained in
RemoteRssSourceUrl, to stop the Windows Phone emulator and other caching
entities from caching the remote request, so you always get the freshest
contents when working in the Visual IDE/debugger. See the notes in code
for further details.