Remote XML Items Files
Added ability to use an XML items file located on a remote server. It
works exactly the same as the LocalItemsFileService, except the rather
than using a local file, users can host their file somewhere else and
load the content in. This allows for Azure-like updating, without the
need for an Azure account.