Updated App Builder
The app builder now has the ability to change the wallpaper for Windows
Phone and Windows 8.

The app is also now created on a background thread to keep the UI from
freezing up. A Progress Bar displays the progress as the app is
generated, which is useful for slower machines.