Add parsing phone #'s from Item.Description
The Webview control in WP8 does not afford us the built-in luxury of
phone # parsing that Windows 8 does, so I have created a shim to inject
JS to detect and create a button for phone numbers matching (xxx)
xxx-xxxx or xxx-xxx-xxxx in Item.Description which when clicked will
call out to C# to make a phone call.  Note that this will only work on a
physical device.