Add support for remote AppSettings
You can now host a JSON representation of AppSettings on a remote
server.  This means your entire app can be configured and updated via a
remote json file.

To ensure the JSON schema is in sync with your published version of
XPCK, a new project has been included "AppSettingsSerializer" which
makes a reference to the "XPlatformCloudKit.PCL" project.  This
guarantees the JSON output of the AppSettingsSerializer matches the
structure of AppSettings.cs in the PCL project.

To use this new feature, run the included AppSettingsSerializer project
and host the result of AppSettings.json on a remote server.  Next, set
EnableRemoteAppSettings to true.
Finally, point to the remote Url which contains your AppSetting.json
output in RemoteAppSettingsService.