Track UrlSource in Item Model
By tracking the UrlSource in the Item Model we can implement custom
logic around the Url and Type in addition to the previously available
Group property.  For example, in ItemsShowcaseViewDataTemplateSelector
we can detect if the Url originates from a version of the Youtube API
and apply a WideTemplate to those items in the ItemsShowcaseView.