Added support for purl.org content tag in rss if present
Purl.org is attempting to create a standard for dealing with dynamic and
changing web content. They have defined an extension to RSS that
provides a more rich rendering of the content in an RSS feed. This tag
is supported by a number of blog engines and, when present, should be
used instead for the description field.