Fix issue where Youtube plays when not in view
By removing the Loaded event from the WebView, we make it eligible for
garbage collection.  We can collect it when the FlipView's selection
changes.  Note: the FlipView holds the next and previous in memory so
they aren't collected.  Thus you can play two videos next to each other
in a group at the same time.  This is potentially desirable.