Refactor to shorter namespaces in Shared Project
It was noticed that when building an .appx package on one of the
Universal Project platforms that the default filepath was usually too
long.  As a result all namespaces and Universal project folders have
been shortened by removing " .Universal." and following the previous
naming convention for project names and appending a "1".